frontend/requirements.txt
flask 
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-flask 
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-proto
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-requests
opentelemetry.instrumentation.logging
